Implemented connection to the RPi WiFi AP programatically via the <connect> button
