+++ BLECar/app/build.gradle
@@ -32,7 +32,4 @@ dependencies {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
     compile 'io.github.controlwear:virtualjoystick:1.7.2'
-    compile "com.polidea.rxandroidble:rxandroidble:1.4.3"
-    implementation 'com.trello:rxlifecycle:1.0'
-    implementation 'com.trello:rxlifecycle-components:1.0'
 }
