+++ BLECar/app/src/main/java/vendetta/blecar/WiFiStateEnum.java
@@ -1,0 +1,11 @@
+package vendetta.blecar;
+
+/**
+ * Created by Vendetta on 29-Apr-18.
+ */
+
+public enum WiFiStateEnum {
+    CONNECTING,
+    CONNECTED,
+    DISCONNECTED
+}
